+++ Riot/Modules/Authentication/AuthenticationViewController.h
@@ -1,6 +1,7 @@
 /*
  Copyright 2015 OpenMarket Ltd
  Copyright 2017 Vector Creations Ltd
+ Copyright 2020 New Vector Ltd
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -17,7 +18,13 @@
 
 #import <MatrixKit/MatrixKit.h>
 
+@protocol AuthenticationViewControllerDelegate;
+
+
 @interface AuthenticationViewController : MXKAuthenticationViewController <MXKAuthenticationViewControllerDelegate>
+
+// MXKAuthenticationViewController has already a `delegate` member
+@property (nonatomic, weak) id<AuthenticationViewControllerDelegate> authVCDelegate;
 
 @property (weak, nonatomic) IBOutlet UIView *navigationBackView;
 @property (weak, nonatomic) IBOutlet UINavigationBar *navigationBar;
@@ -50,3 +57,9 @@
 
 @end
 
+
+@protocol AuthenticationViewControllerDelegate <NSObject>
+
+- (void)authenticationViewControllerDidDismiss:(AuthenticationViewController *)authenticationViewController;
+
+@end;