+++ CHANGES.rst
@@ -20,7 +20,7 @@ Changes to be released in next version
  * 
 
 Others
- * 
+ * buildRelease.sh: Pass a `git_tag` parameter to fastlane because fastlane `git_branch` method can fail.
 
 Changes in 1.0.10 (2020-09-08)
 =================================================