+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changes to be released in next version
 
 🙌 Improvements
  * Room: Collapse state messages on room creation (#3629). 
- * AuthVC: Make force PIN working for registration as well
+ * AuthVC: Make force PIN working for registration as well.
+ * AppDelegate: Do not show incoming key verification requests while authenticating.
 
 🐛 Bugfix
  * AuthVC: Fix PIN setup that broke cross-signing bootstrap.