+++ CHANGES.rst
@@ -6,9 +6,12 @@ Changes to be released in next version
 
 🙌 Improvements
  * Room: Collapse state messages on room creation (#3629). 
+ * AuthVC: Make force PIN working for registration as well.
+ * AppDelegate: Do not show incoming key verification requests while authenticating.
 
 🐛 Bugfix
- * 
+ * AuthVC: Fix PIN setup that broke cross-signing bootstrap.
+ * Loading animation: Fix the bug where, after authentication, the animation disappeared too early and made auth screen flashed.
 
 ⚠️ API Changes
  *